results output

+++ test_framework.py
@@ -12,6 +12,7 @@ from datetime import datetime
 import subprocess
 import evaluation
 import global_settings
+import os.path
 
 
 aux_verb = ['was', 'is', 'become']
@@ -24,7 +25,10 @@ ambiverse = True
 global_settings.init()
 # data_source = 'ug_data/all_'
 data_source = 'main_data/'
-output_data = 'output_data/'
+# data_source = 'ug_final/'
+if not os.path.exists('output_data'+str(time.time())+'/'):
+    os.makedirs('output_data'+str(time.time())+'/')
+    output_data = 'output_data'+str(time.time())+'/'
 
 def json_serial(obj):
     """JSON serializer for objects not serializable by default json code"""
@@ -66,6 +70,7 @@ def fact_checker(sentence_lis, id_list):
     new_triple_flag = 0
     new_predicate_flag = 0
     output_linkprediction = {}
+    output_training_data = {}
     for i in range(0, 1):
         for n, ne in enumerate(ne_s):
             sent_id = id_list[n]
@@ -84,9 +89,12 @@ def fact_checker(sentence_lis, id_list):
                 print "NER: "+str(ent_dict)
                 # sys.exit(0)
                 # resources, ent_size, date_labels, raw_resources = fact_check.resource_extractor(ent)
+                fact_check.get_labels(ent)
+                # sys.exit(0)
             res_time = time.time()
             new_triple_flag,triple_dict = triples_extractor(sent_id, sentence_lis[n],ne, new_triple_flag)
             print "Relation Triples: "+str(triple_dict)
+            # sys.exit(0)
             # print entity_matched
             #     resources = entity_matched
             relation = []
@@ -97,16 +105,25 @@ def fact_checker(sentence_lis, id_list):
             print "=================="
             pprint.pprint(resource_text)
             resource_ids = fact_check.get_resource_id(resource_text)
+            print resource_ids
             precision_ent, recall_ent, entity_matched = evaluation.precision_recall_entities(sent_id, resource_text)
-            precision_res = sum(precision_ent)/len(precision_ent)
-            recall_res = sum(recall_ent)/len(recall_ent)
+            if precision_ent:
+                precision_res = sum(precision_ent)/len(precision_ent)
+            else:
+                precision_res = 0
+            if recall_ent:
+                recall_res = sum(recall_ent)/len(recall_ent)
+            else:
+                recall_res = 0
             relation_ent = fact_check.relation_extractor_triples(resource_text, triple_dict, relation)
             # print relation_ent
             if not relation_ent:
                 sentence_list = [word_tokenize(sent) for sent in sentence_lis]
                 ne_s, pos_s, dep_s = fact_check.st_tagger(sentence_list)
                 verb_entity = fact_check.verb_entity_matcher(dep_s)
-                relation_ent, rel_count = fact_check.relation_extractor_all(resources, verb_entity[n])
+                relation_ent, rel_count = fact_check.relation_extractor_all(resource_text, verb_entity[n])
+                # print relation_ent
+                # sys.exit(0)
             # print "Precision & Recall for Resource Extractor"
             # print "-----------------------------------------"
             relations = fact_check.relation_processor(relation_ent)
@@ -124,7 +141,7 @@ def fact_checker(sentence_lis, id_list):
                 true_pos_rel, retrived_rels, ex_rels = evaluation.precision_recall_relations(sent_id, relations)
                 print true_pos_rel, retrived_rels, ex_rels
                 true_pos_ent, retrieved_ents, ex_ent_all = evaluation.precision_recall_ent_match(sent_id, relations)
-                print ex_ent_all
+                # print ex_ent_all
                 print '\n'
                 print "Precision & Recall for Entities"
                 print "--------------------------------"
@@ -136,43 +153,52 @@ def fact_checker(sentence_lis, id_list):
                 precision_rel, recall_rel = evaluation.precision_recall(true_pos_rel, retrived_rels, ex_rels)
                 print "Relations: Precision: " + str(precision_rel), "Recall: " + str(recall_rel)
                 precision_recall_stats[sent_id] = [precision_res, recall_res, precision_rel, recall_rel, precision_ent_out, recall_ent_out]
-                if KG_Miner:
-                    print "Using KG_Miner"
-                    entity_type_ontology, entity_type_resource = KG_Miner_Extension.entity_type_extractor(resources,\
-                                                                                                          triple_dict)
-                    print "Type of Entities"
-                    pprint.pprint(entity_type_ontology)
-                    # pprint.pprint(entity_type_resource)
-                    # sys.exit(0)
-                    # resource_type_set_ranked, resource_threshold_ranked = KG_Miner_Extension.entity_type_ranker(entity_type_resource, ent_dict, triple_dict)
-                    # ontology_type_set_ranked, ontology_threshold_ranked = KG_Miner_Extension.entity_type_ranker\
-                    #     (entity_type_ontology, ent_dict, triple_dict)
-                    # pprint.pprint(resource_type_set_ranked)
-                    # pprint.pprint(ontology_type_set_ranked)
-                    # pprint.pprint(ontology_threshold_ranked)
-                    # sys.exit(0)
-                    if sent_id in possible_predicate.keys():
-                        possible_predicate_set = possible_predicate[sent_id]
-                    else:
-                        possible_predicate_set = KG_Miner_Extension.possible_predicate_type(entity_type_ontology, triple_dict, resource_ids)
-                        possible_predicate[sent_id] = possible_predicate_set
-                        new_predicate_flag=1
-                    # print possible_predicate_set
-                    # possible_predicate_set_ranked, possible_predicate_set_threshold = KG_Miner_Extension.predicate_ranker(possible_predicate_set,triple_dict)
-                    # print possible_predicate_set_ranked
-                    # print possible_predicate_set_threshold
-                    # sys.exit(0)
-                    try:
-                        os.remove('KG_Miner_data/poi.csv')
-                        os.remove('KG_Miner_data/predicate_probability.csv')
-                    except:
-                        pass
-                    predicate_results = KG_Miner_Extension.get_training_set(possible_predicate_set, \
-                                                                            entity_type_resource, entity_type_ontology, \
-                                                                            triple_dict, resource_ids)
-                    output_linkprediction[id_list[n]] = predicate_results
             else:
                 precision_recall_stats[sent_id] = [0, 0, 0, 0, 0, 0]
+            if KG_Miner:
+                print "Using KG_Miner"
+                entity_type_ontology, entity_type_resource = KG_Miner_Extension.entity_type_extractor(resources,\
+                                                                                                      triple_dict)
+                training_entity_type, training_resource_type = KG_Miner_Extension.training_entity_type(resources, triple_dict)
+
+                print "Type of Entities"
+                pprint.pprint(entity_type_ontology)
+                pprint.pprint(entity_type_resource)
+                # sys.exit(0)
+                # sys.exit(0)
+                # resource_type_set_ranked, resource_threshold_ranked = KG_Miner_Extension.entity_type_ranker(entity_type_resource, ent_dict, triple_dict)
+                # ontology_type_set_ranked, ontology_threshold_ranked = KG_Miner_Extension.entity_type_ranker\
+                #     (entity_type_ontology, ent_dict, triple_dict)
+                # pprint.pprint(resource_type_set_ranked)
+                # pprint.pprint(ontology_type_set_ranked)
+                # pprint.pprint(ontology_threshold_ranked)
+                # sys.exit(0)
+                if sent_id in possible_predicate.keys():
+                    possible_predicate_set = possible_predicate[sent_id]
+                else:
+                    possible_predicate_set = KG_Miner_Extension.possible_predicate_type(entity_type_ontology, triple_dict, resource_ids)
+                    possible_predicate[sent_id] = possible_predicate_set
+                    new_predicate_flag=1
+                # print possible_predicate_set
+                # possible_predicate_set_ranked, possible_predicate_set_threshold = KG_Miner_Extension.predicate_ranker(possible_predicate_set,triple_dict)
+                # print possible_predicate_set_ranked
+                # print possible_predicate_set_threshold
+                # sys.exit(0)
+                try:
+                    os.remove('KG_Miner_data/poi.csv')
+                    os.remove('KG_Miner_data/predicate_probability.csv')
+                except:
+                    pass
+                # print entity_type_ontology
+                # print training_entity_type
+                # print training_resource_type
+                # sys.exit(0)
+                predicate_results, training_data_set = KG_Miner_Extension.get_training_set(possible_predicate_set, \
+                                                                        training_resource_type, training_entity_type, \
+                                                                        triple_dict, resource_ids)
+                output_linkprediction[id_list[n]] = predicate_results
+                output_training_data[id_list[n]] = training_data_set
+            
             execution_time = time.time() - res_time
             print "Execution Time: " + str(round(execution_time, 2))
             print "================================================="
@@ -182,6 +208,9 @@ def fact_checker(sentence_lis, id_list):
             with open(output_data+'/link_prediction.json', 'w') as fp:
                 json.dump(output_linkprediction, fp, default=json_serial)
 
+            with open(output_data+'/training_data.json', 'w') as fp:
+                json.dump(output_training_data, fp, default=json_serial)
+
             with open(output_data+'/output_relations.json', 'w') as fp:
                 json.dump(output_realations, fp, default=json_serial)
         print output_linkprediction
@@ -190,19 +219,22 @@ def fact_checker(sentence_lis, id_list):
     
 
     if new_triple_flag == 1:
-        os.remove(data_source+'triples_raw.json')
+        if os.path.isfile(data_source + 'triples_raw.json'):
+            os.remove(data_source+'triples_raw.json')
         with open(data_source+'/triples_raw.json', 'w') as fp:
             json.dump(file_triples, fp, default=json_serial)
 
     if global_settings.new_ambi_query==1:
         first_query = ambiverse_api.first_ambiverse()
-        os.remove(data_source+'ambiverse_resources.json')
+        if os.path.isfile(data_source+'ambiverse_resources.json'):
+            os.remove(data_source+'ambiverse_resources.json')
         with open(data_source+'ambiverse_resources.json', 'w') as fp:
             json.dump(first_query, fp, default=json_serial)
 
 
     if new_predicate_flag == 1:
-        os.remove(data_source+'possible_predicate.json')
+        if os.path.isfile(data_source+'possible_predicate.json'):
+            os.remove(data_source+'possible_predicate.json')
         with open(data_source+'/possible_predicate.json', 'w') as fp:
             json.dump(possible_predicate, fp, default=json_serial)
 
@@ -226,9 +258,11 @@ def fact_checker(sentence_lis, id_list):
         print "average",vals_avg
 
 
-
-with open(data_source+'triples_raw.json') as json_data:
-    file_triples = json.load(json_data)
+if os.path.isfile(data_source+'triples_raw.json'):
+    with open(data_source+'triples_raw.json') as json_data:
+        file_triples = json.load(json_data)
+else:
+    file_triples = {"a":"b"}
 
 with open('nodes_id.json') as json_data:
     global_settings.nodes_id = json.load(json_data)
@@ -237,14 +271,22 @@ with open('nodes_id.json') as json_data:
 with open('edge_types_id.json') as json_data:
     global_settings.edge_id = json.load(json_data)
 
-with open(data_source+'ambiverse_resources.json') as json_data:
-    ambiverse_api.ambiverse_resources = json.load(json_data)
 
-with open(data_source+'possible_predicate.json') as json_data:
-    possible_predicate = json.load(json_data)
+if os.path.isfile(data_source+'ambiverse_resources.json'):
+    with open(data_source+'ambiverse_resources.json') as json_data:
+        ambiverse_api.ambiverse_resources = json.load(json_data)
+else:
+    ambiverse_api.ambiverse_resources = {"a":"b"}
 
 
-with open(data_source+'sentences.csv') as f:
+if os.path.isfile(data_source+'possible_predicate.json'):
+    with open(data_source+'possible_predicate.json') as json_data:
+        possible_predicate = json.load(json_data)
+else:
+    possible_predicate = {"a":"b"}
+
+
+with open(data_source+'sentences1.csv') as f:
     reader = csv.DictReader(f)
     sentences_list = []
     id_list = []
